refactor: extract diagram export helpers
