+++ README.md
@@ -1,4 +1,4 @@
-version: 0.2.3
+version: 0.2.4
 Author: Miguel Marina <karel.capek.robotics@gmail.com> - [LinkedIn](https://www.linkedin.com/in/progman32/)
 # AutoML
 
@@ -1636,6 +1636,7 @@ and run the build again if you hit this issue.
 
 
 ## Version History
+- 0.2.4 - Extracted diagram export helpers into dedicated sub-app.
 - 0.2.3 - Moved malfunction, fault, hazard, and insufficiency handlers into risk assessment sub-app.
 - 0.2.2 - Moved risk assessment helpers into dedicated sub-app.
 - 0.2.1 - Extracted review diff and version functions into ReviewManager.
