+++ mainappsrc/use_case_diagram_subapp.py
@@ -21,3 +21,8 @@ class UseCaseDiagramSubApp:
         self.app.diagram_tabs[diag.diag_id] = tab
         UseCaseDiagramWindow(tab, self.app, diagram_id=diag.diag_id)
         self.app.refresh_all()
+
+    # ------------------------------------------------------------------
+    def create_export_window(self, parent, diagram):
+        """Return a window instance for exporting *diagram*."""
+        return UseCaseDiagramWindow(parent, self.app, diagram_id=diagram.diag_id)
